fix 1

+++ Kinh_Doanh_Khoa_Hoc_Truc_Tuyen_Domain/Entities/OrderDetail.cs
@@ -14,9 +14,9 @@ namespace Kinh_Doanh_Khoa_Hoc_Truc_Tuyen_Domain.Entities
 
         public Guid ActiveCourseId { set; get; }
 
-        public decimal? Price { set; get; }
+        public long? Price { set; get; }
 
-        public decimal? PromotionPrice { set; get; }
+        public long? PromotionPrice { set; get; }
 
         public Order Order { set; get; }
